pyproject.toml
[project]
name = "damn-vulnerable-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.10",
    "mcp[cli]>=1.26.0",
    "python-dotenv>=1.2.1",
    "starlette>=0.52.1",
    "uvicorn>=0.41.0",
]
